fixed express handlebars missing functions
