+++ .env.Example
@@ -1,5 +1,7 @@
-DB_NAME=job_board_db
-DB_USER=root
-DB_PASSWORD=
+# local mysql credentials
+DB_NAME=
+DB_USER=
+DB_PW=
 
+# secret for signing cookies
 SECRET=keyboardcat
