+++ .gitignore
@@ -1,6 +1,3 @@
-node_modules
 .DS_Store
+node_modules
 .env
-package-lock.json
-package.json
-test.html
