+++ controllers/api/users-router.js
@@ -2,11 +2,9 @@ const { User } = require("../../models");
 const router = require("express").Router();
 
 router.post("/", async (req, res) => {
+  const { username, password } = req.body;
   try {
-    const user = await User.create({
-      email: req.body.email, 
-      password: req.body.password
-    });
+    const user = await User.create(req.body, { username, password });
     req.session.isLoggedIn = true;
     req.session.userId = user.id;
     req.session.save(() => res.json({ id: user.id }));
@@ -16,7 +14,26 @@ router.post("/", async (req, res) => {
   }
 });
 
-// Does this really work?
+router.post("/login", async (req, res) => {
+  const { username, password } = req.body;
+  try {
+    const user = await User.findOne({ where: { username } });
+    if (!user) {
+      throw new Error("User not found.");
+    }
+    const isValidPassword = await user.checkPassword(password);
+    if (!isValidPassword) {
+      throw new Error("Invalid password");
+    }
+    req.session.isLoggedIn = true;
+    req.session.userId = user.id;
+    req.session.save(() => res.json({ id: user.id }));
+  } catch (error) {
+    console.error(error);
+    res.status(400).json({ message: "Invalid username or password." });
+  }
+});
+
 router.post("/logout", (req, res) => {
   req.session.destroy((err) => {
     if (err) {
