+++ controllers/index.js
@@ -1,8 +1,6 @@
-const router = require('express').Router();
-const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes');
-
-router.use('/', homeRoutes);
-router.use('/api', apiRoutes);
-
+const router = require("express").Router();
+const homeRouter = require("./home-router");
+const apiRouter = require("./api");
+router.use("/", homeRouter);
+router.use("/api", apiRouter);
 module.exports = router;
