+++ package.json
@@ -1,9 +1,8 @@
 {
-  "name": "help-for-hire",
+  "name": "express-hbars-boilerplate",
   "version": "1.0.0",
-  "description": "",
+  "description": "This boilerplate may be used as a starting point for building a full-stack app with Handlebars, Express, and Sequelize/MySQL.",
   "private": true,
-  "main": "server.js",
   "scripts": {
     "start": "node server.js",
     "test": "npm run lint",
@@ -11,31 +10,25 @@
     "lintfix": "eslint --fix .",
     "format": "prettier --write .",
     "watch": "nodemon server.js",
-    "seeds": "node seeds"
+    "seed": "node seed"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/MaestroLuu/Help-for-Hire.git"
-  },
-  "author": "David Luu",
+  "keywords": [],
+  "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/MaestroLuu/Help-for-Hire"
-  },
   "dependencies": {
     "bcrypt": "^5.0.1",
-    "connect-session-sequelize": "^7.1.2",
-    "dotenv": "^16.0.0",
-    "express": "^4.17.2",
-    "express-handlebars": "^6.0.2",
+    "connect-session-sequelize": "^7.1.1",
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "express-handlebars": "^5.3.2",
     "express-session": "^1.17.2",
-    "mysql2": "^2.3.3",
-    "sequelize": "^6.16.0"
+    "mysql2": "^2.2.5",
+    "sequelize": "^6.6.5"
   },
   "devDependencies": {
     "eslint": "^8.8.0",
     "eslint-config-prettier": "^8.3.0",
     "nodemon": "^2.0.15",
-    "prettier": "^2.5.1"
+    "prettier": "^2.3.2"
   }
 }
