--- test.html
@@ -1,44 +1,0 @@
-<html lang='en'>
-
-<head>
-    <meta charset='UTF-8' />
-    <meta http-equiv='X-UA-Compatible' content='IE=edge' />
-    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
-    <title>{{title}}</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <link rel='stylesheet' href="./public/css/style.css">
-</head>
-
-<body>
-    <header>
-        <nav class="d-flex col-12 justify-content-between p-2 bg-secondary position-fixed">
-            <div>
-                <button class='btn btn-about-us'>About Us</button>
-            </div>
-            <div>
-                {{#if isLoggedIn}}
-                <button class='btn btn-logout'>Log Out</button>
-                {{else}}
-                <button class='btn btn-login'>Log In</button>
-                <button class='btn btn-signup'>Sign Up</button>
-                {{/if}}
-            </div>
-        </nav>
-    </header>
-    <h1 class="pt-5 text-center">Signup</h1>
-    <form class='signup-form text-center'>
-      <label for='username'>Username</label><br />
-      <input type='text' name='username' id='username' /><br />
-      <label for='password'>Password</label><br />
-      <input type='password' name='password' id='password' /><br />
-      <label for='confirm-password'>Confirm Password</label><br />
-      <input type='password' name='confirm-password' id='confirm-password' /><br />
-      <button class='btn m-5' type='submit'>Submit</button>
-    </form>
-    <p class="text-center">Already have an account? <a href='/login'>Login Here</a></p>
-    <script src='/js/signup.js'></script>
-    <script src='./public/js/main.js'></script>
-</body>
-
-</html>
