+++ utils/helpers.js
@@ -1,6 +1,5 @@
 module.exports = {
-    // add helper functions for handlebars here
-    // Example:
-    // json: object => JSON.stringify(object, null, 4),
-  };
-  
+  // add helper functions for handlebars here
+  // Example:
+  // json: object => JSON.stringify(object, null, 4),
+};
