+++ views/home.handlebars
@@ -1,13 +1,13 @@
-    <div class="container-fluid text-center">
-        <h1 class="pt-5">Home</h1>
-        <h2>Welcome, {{user.username}}!</h2>
-        <p class="border-top border-dark fs-3 align-center">Please select an option</p>
-        <div class="align-center">
-            <div class="m-5">
-                <button class='btn btn-seeking text-primary bg-light'>Seeking</button>
-            </div>
-            <div class="m-5">
-                <button class='btn btn-hiring text-success bg-light'>Hiring</button>
-            </div>
+<div class="container-fluid text-center">
+    <h1 class="pt-5">Home</h1>
+    <h2>Welcome, {{user.username}}!</h2>
+    <p class="border-top border-dark fs-3 align-center">Please select an option</p>
+    <div class="align-center">
+        <div class="m-5">
+            <button class='btn btn-seeking text-primary bg-light'>Seeking</button>
+        </div>
+        <div class="m-5">
+            <button class='btn btn-hiring text-success bg-light'>Hiring</button>
         </div>
     </div>
+</div>
