+++ views/layouts/main.handlebars
@@ -1,33 +1,33 @@
 <html lang='en'>
 
 <head>
-    <meta charset='UTF-8' />
-    <meta http-equiv='X-UA-Compatible' content='IE=edge' />
-    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
-    <title>{{title}}</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <link rel='stylesheet' href="/css/style.css">
+  <meta charset='UTF-8' />
+  <meta http-equiv='X-UA-Compatible' content='IE=edge' />
+  <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+  <title>{{title}}</title>
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
+      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+  <link rel='stylesheet' href="/css/style.css">
 </head>
 
 <body>
-    <header>
-        <nav class="d-flex col-12 justify-content-between p-2 bg-secondary position-fixed">
-            <div>
-                <button class='btn btn-about-us'>About Us</button>
-            </div>
-            <div>
-                {{#if isLoggedIn}}
-                <button class='btn btn-logout'>Log Out</button>
-                {{else}}
-                <button class='btn btn-login'>Log In</button>
-                <button class='btn btn-signup'>Sign Up</button>
-                {{/if}}
-            </div>
-        </nav>
-    </header>
-    {{{body}}}
-    <script src='/js/main.js'></script>
+  <header>
+    <nav class="d-flex col-12 justify-content-between p-2 bg-secondary position-fixed">
+      <div>
+        <button class='btn btn-about-us'>About Us</button>
+      </div>
+      <div>
+        {{#if isLoggedIn}}
+        <button class='btn btn-logout'>Log Out</button>
+        {{else}}
+        <button class='btn btn-login'>Log In</button>
+        <button class='btn btn-signup'>Sign Up</button>
+        {{/if}}
+      </div>
+    </nav>
+  </header>
+  {{{body}}}
+  <script src='/js/main.js'></script>
 </body>
 
 </html>
