+++ views/login.handlebars
@@ -1,10 +1,10 @@
-    <h1 class="pt-5 text-center">Login</h1>
-    <form class='login-form text-center'>
-      <label for='email'>Email</label><br>
-      <input type='text' name='email' id='email'><br>
-      <label for='password'>Password</label><br>
-      <input type='password' name='password' id='password'><br>
-      <button class='btn m-5' type='submit'>Submit</button>
-    </form>
-    <p class="text-center">Don't have an account yet? <a href='/signup'>Sign Up Here</a></p>
-    <script src='/js/login.js'></script>
+<h1>Login</h1>
+<form class='login-form'>
+  <label for='username'>Username</label><br />
+  <input type='text' name='username' id='username' /><br />
+  <label for='password'>Password</label><br />
+  <input type='password' name='password' id='password' /><br />
+  <button class='btn' type='submit'>Submit</button>
+</form>
+<p>Don't have an account yet? <a href='/signup'>Sign Up Here</a></p>
+<script src='/js/login.js'></script>
