+++ views/signup.handlebars
@@ -1,12 +1,12 @@
-<h1 class="pt-5 text-center">Signup</h1>
-<form class='signup-form text-center'>
-  <label for='email'>Email</label><br />
-  <input type='text' name='email' id='email' /><br />
+<h1>Signup</h1>
+<form class='signup-form'>
+  <label for='username'>Username</label><br />
+  <input type='text' name='username' id='username' /><br />
   <label for='password'>Password</label><br />
   <input type='password' name='password' id='password' /><br />
   <label for='confirm-password'>Confirm Password</label><br />
   <input type='password' name='confirm-password' id='confirm-password' /><br />
-  <button class='btn m-5' type='submit'>Submit</button>
+  <button class='btn' type='submit'>Submit</button>
 </form>
 <p>Already have an account? <a href='/login'>Login Here</a></p>
 <script src='/js/signup.js'></script>
